Initial Vagrant Check-in
Getting the homepage working with vagrant (with a bit of extra configuration still). Getting closer to a better setup process which is also closer to continuous integration.

Configuration is migrating from a settings file in the app to a .env file for local configuration and environment variables (or an equivalent .env file) for stage and production environments.

Right now DB stuff is more a 1:1 mapping to get things going but this will change. Metadata and Feature Def tables should be getting pulled out. Additionally, several databases are made on intialization, but ideally we would only make the one (or more) needed for the current running implementation. From there, a switch in configuration would likely be easiest to do with a `vagrant provision` and/or changing of settings in the .env file.