Data Upload URLs
Not hooked up to a front end, but has a test page at `/data/test`.

Uploads to cloud storage if configured, otherwise uploads locally to the vagrant machine into an "uploads/" directory not contained in the project (local to the vagrant machine and not shared).

Adds the following Environment Variables:

 * `USE_CLOUD_STORAGE` - Flag to use cloud storage for file upload. **DEFAULT: `False`. Set to `True` on production environments**
 * `GCLOUD_PROJECT_NAME` - Name of the project that will be uploaded to
 * `GOOGLE_APPLICATION_CREDENTIALS` - Path to privatekey.json file with GCloud Credentials
 * `GOOGLE_STORAGE_BUCKET` - Bucket to place uploaded files in by default
 * `MEDIA_FOLDER` - Folder to place uploaded files in. Works if uploading to the cloud or locally **DEFAULT: `uploads/`**