Fix double display bug upon sharing with multiple people
Combining querysets in Django is far harder than it needed to be. All you need to do when you have two sets of the same model is a simple `UNION`. But they instead try to get fancy with their `OR`'s and end up messing up the queries anyway. So now instead of a simple set union, I also have to do a distinct. Because if you don't, you end up with lots of dupes. One for each person you shared with.

At least it's fixed now.