Adding another migration for workbooks
I was trying to avoid adding this one because it relies on projects which isn't checked in. I didn't check in that migration because it gets complicated from there because of other dependencies it has. Basically, at that point, we would need to wholesale checkin migrations. Which we should probably do anyway, but is a bit much for now, for this.